services/pathfinding-engine/requirements.txt
fastapi
uvicorn
networkx
websockets
requests
python-socketio
